+ long user id added to model and new intent for view pager added and task repository updated with user id and places that used task repository fixed.

+++ app/src/main/java/project/com/maktab/hw_6/model/task/Task.java
@@ -10,13 +10,13 @@ public class Task {
     private Date mDate;
     private UUID mID;
     private String mTaskType;
-    private int mUserID;
+    private long mUserID;
 
-    public int getUserID() {
+    public long getUserID() {
         return mUserID;
     }
 
-    public void setUserID(int userID) {
+    public void setUserID(long userID) {
         mUserID = userID;
     }
